+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ### v0.1.38
 - Script : Add `.script_name` as an alias for `.name`
 - Stage : Add safety checking to *Card-related methods + error for invalid string identifiers
-- TOML : Add `new_config`
+- TOML.`new` : Add optional `directory` argument
 - Util : Add `get_namespace`
 
 ### v0.1.37